Non-volatile KV store on nvs partition
This implements the following new extensions that get/set arbitrary
lbm values on the nvs partition using LBM flat values.

eeprom-* can use up to 256*32 bytes of data on the nvs partition,
leaving 16k for KV data (and more if eeprom is not fully utilized).

This is useful if you want to use the QML partition for something else,
for example QML.

API:
kv is for key-value

(kv-get key) ; return value, nil if not set
(kv-get key default) ;return value, default if not set
(kv-set key value) ; set value
(kv-erase key) ; erase value
(kv-list) ;  list KV keys

Since this is now sharing storage with eeprom-*, it is useful to be
able to erase eeprom varables to free up space, this can be done with:

(eeprom-erase addr)