feat: Merge "date/" and "no_date/" log folders
Reorganizes how the log files are named so that they always contain a
shared incrementing index and places them in the same folder "log_can/".

Example for how the log names would be set:
1. If date is known:
   `log_<index>_2025-04-07_17-57-00.csv`
2. If date is not known:
   `log_<index>.csv`

This is very useful when you have an unreliable GPS connection, where
some of the logs will be generated when you have GPS (and therefore a
known time), and some when you don't. You may even have some where the
GPS connection was half functional, so that you got a file with a date
but the time fields set to "00:00:00". This makes it very hard to
confidently identify which log was the latest one.

Now that the logs share the same incrementing index they will always
show up in the order they were created in.