Fix `wifi-connect` breaking if called while connecting
The `wifi-connect` LBM extension has a bug where it will return `nil`
incorrectly when called while the system is in the connecting state,
despite the connection succeeding (assuming that the credentials were
correct).
This was fixed by adding a workaround to `comm_wifi_change_network`,
which you can read about in the comment around it.

Also fixed a bug in `wifi-connect`'s implementation which would cause it
to lock the wifi extensions if `comm_wifi_change_network` failed,
without actually blocking the current thread!