Add `wifi-connect-last` and `'no-wait` flags; Fix precheck conditions
Adds the new `wifi-connect-last` wifi extension which tries to connect
to the network from the last call to `wifi-connect`---if it has been
called previously---or from the VESC config.

Also adds the new `'no-wait` flag which you can pass to `wifi-connect`
and `wifi-connect-last`. This makes them return immediately, without
blocking until the result of the connection is known.

The target usecase for both of these features is to be able to re-enable
auto-reconnection after having called `wifi-disconnect`, but only to
disconnect temporarily e.g. to be able to call `wifi-scan-networks`
(which requires that we aren't trying to connect to a network
currently). In this case you may not care about knowing how the
connection went, since you just want it to continue to attempt
reconnecting if not connected as it did before.

I also updated the precheck conditions to make them more useful. The
important change was that the tcp extensions are no longer blocked by
the wifi extentions being called by another thread. There wasn't really
a reason to have that restriction in the first case, since they are
either not blocking, or use separate C-threads, meaning they won't
override any stored context IDs.

The documentation has been updated to match these changes. I also added
explicit notes to the extensions that can't be called asynchronously.